rename files of controllers, routes and models folders
