+++ routes/register.js
@@ -1,4 +1,4 @@
-const handlePost = require("../controllers/register.controller");
+const handlePost = require("../controllers/register");
 
 module.exports = app => {
   app.post("/register", handlePost);
